Move run_standalone to VCRIntegrationTest
